Largely Complete Inventory System
The inventory system works now. You should handle it through the Inventory Manager. The Inventory Manage script can be put on anything, but you do need to have a UI pannel inside a UI Canvas as well as an EventSystem (just don't delete the existing EventSystem and it'll be fine).

addItem(Item) or addItem(itemId, numberToAdd) are the two ways to add items to the player's inventory.

removeItem(itemId, numberToRemove) is how you should take stuff out

renameItem, changeItemDescription, changeItemSprite all take the itemId and the newValue, and change the inventory data according to their function name.

swapItems(idA, idB) will swap the position of two items in the inventory. Mostly in case we ever decide to do some sort of dragging inventory items thing.

These all update both the backend and the visual front end at once.

In addition to the inventory system, I set up an item management system. It's just something to keep all the data about items in one place, and theoretically easily set the appropriate default sprites for things. This could be expanded to maybe take in some sort of text file (probably a csv file or a json), but thats something for later, and would probably only make sense if you end up with a large number of items.