Figured out how to do beams of light
I'm an idiot...
Of course spotlights that are barely in front of the image and perpendicular will create beams...
Anyway, darkened the ambient lighting so the only lighting comes from unity lights placed in the scene.
Set up a flashlight that can stick to a source and which will follow a target.