Added Swapping inventories
The inventory manager will now keep track of multiple inventories.
Only one can be used/displayed at a time, but one can swap between them as much as you might want.

Add new Inventories with addInventory(string newName)
(it will return false if there is already an inventory with that name)

Change the active inventory with changeInventories(string otherInvName)
(it will return false if there is no inventory with the name provided)

Delete unneeded inventories with removeInventory(string nameToRemove)
(it will return false if the specified inventory did not exist)

Look up the name of the current inventory with getNameOfActiveInv()