Added the world interactions
Add the UseItem Script (or better yet a decendent) to an object (I believe it needs a rigid body2d and a 2d collider to work) and alter the Start and useItem(Item item) functions to suit the exact situation.

InventoryManager will look for when you've clicked on an object with the UseItem class (or decendent) and run the useItem function at that time.

FYI, forgot to mention earlier:
Test inventory adding by pressing Delete
Test inventory swapping with Space (after adding an inventory with Delete).